Fix error message typo in getTvShows function and update error message in getTopRatedMovies function
This commit corrects a typo in the error message of the `getTvShows` function, changing "Unable to tv shows" to "Unable to fetch tv shows". Additionally, it updates the error message in the `getTopRatedMovies` function to provide more clarity, stating "Unable to fetch top rated movies" in case of a non-ok response status.