Add SortDialog component
Created a new component called SortDialog. This component is responsible for displaying a dialog box with sorting options. It takes props such as title, open, onClose, onSortChange, and sortOptions to customize its behavior. The SortDialog component renders a Material-UI Dialog with a title and a list of sorting options provided through the sortOptions prop. Users can select an option, triggering the onSortChange callback function.